// src/app/types/schema.ts
export interface Project extends SanityDocument {
  _type: "project";

  /**
   * seo — `seo`
   *
   *
   */
  seo?: Seo;

  /**
   * Title — `string`
   *
   *
   */
  title?: string;

  /**
   * Slug — `slug`
   *
   * URL based on the title (no space, or char other than a-z-0-9
   */
  slug?: { _type: "slug"; current: string };

  /**
   * look — `string`
   *
   *
   */
  look?: "default" | "sift";

  /**
   * Image Cover — `image`
   *
   *
   */
  imageCover?: {
    _type: "image";
    asset: SanityReference<SanityImageAsset>;
    crop?: SanityImageCrop;
    hotspot?: SanityImageHotspot;
  };

  /**
   * Text — `blockContent`
   *
   *
   */
  text?: BlockContent;

  /**
   * Modules — `array`
   *
   *
   */
  modules?: Array<
    | SanityKeyed<ImageUI>
    | SanityKeyed<TextUI>
    | SanityKeyed<TextImageUI>
    | SanityKeyed<EmbedUI>
  >;

  /**
   * year — `number`
   *
   *
   */
  year?: number;

  /**
   * client — `string`
   *
   *
   */
  client?: string;

  /**
   * industry — `string`
   *
   *
   */
  industry?: string;

  /**
   * location — `string`
   *
   *
   */
  location?: string;

  /**
   * credits — `array`
   *
   *
   */
  credits?: Array<SanityKeyed<Contact>>;

  /**
   * links — `array`
   *
   *
   */
  links?: Array<SanityKeyed<Contact>>;
}

/**
 * Space
 *
 *
 */
export interface Space extends SanityDocument {
  _type: "space";

  /**
   * seo — `seo`
   *
   *
   */
  seo?: Seo;

  /**
   * Title — `string`
   *
   *
   */
  title?: string;

  /**
   * Slug — `slug`
   *
   * URL based on the title (no space, or char other than a-z-0-9
   */
  slug?: { _type: "slug"; current: string };

  /**
   * look — `string`
   *
   *
   */
  look?: "default" | "sift";

  /**
   * Image Cover — `image`
   *
   *
   */
  imageCover?: {
    _type: "image";
    asset: SanityReference<SanityImageAsset>;
    crop?: SanityImageCrop;
    hotspot?: SanityImageHotspot;
  };

  /**
   * Text — `blockContent`
   *
   *
   */
  text?: BlockContent;

  /**
   * Modules — `array`
   *
   *
   */
  modules?: Array<
    | SanityKeyed<ImageUI>
    | SanityKeyed<TextUI>
    | SanityKeyed<TextImageUI>
    | SanityKeyed<EmbedUI>
  >;

  /**
   * year — `number`
   *
   *
   */
  year?: number;

  /**
   * client — `string`
   *
   *
   */
  client?: string;

  /**
   * industry — `string`
   *
   *
   */
  industry?: string;

  /**
   * location — `string`
   *
   *
   */
  location?: string;

  /**
   * credits — `array`
   *
   *
   */
  credits?: Array<SanityKeyed<Contact>>;

  /**
   * links — `array`
   *
   *
   */
  links?: Array<SanityKeyed<Contact>>;
}

/**
 * Tag
 *
 *
 */
export interface Tag extends SanityDocument {
  _type: "tag";

  /**
   * Title — `string`
   *
   *
   */
  title?: string;

  /**
   * Description — `text`
   *
   *
   */
  description?: string;
}

export type Seo = {
  _type: "seo";
  /**
   * Meta title — `string`
   *
   *
   */
  metaTitle?: string;

  /**
   * Meta description — `string`
   *
   * ~155-160 characters
   */
  metaDescription?: string;

  /**
   * Meta image — `image`
   *
   * jpg ou png au format paysage 2:1 (ex: 1200x600px)
   */
  metaImage?: {
    _type: "image";
    asset: SanityReference<SanityImageAsset>;
    crop?: SanityImageCrop;
    hotspot?: SanityImageHotspot;
  };
};
// ...
export type BlockContent = Array<
  | SanityKeyed<SanityBlock>
  | SanityKeyed<{
      _type: "image";
      asset: SanityReference<SanityImageAsset>;
      crop?: SanityImageCrop;
      hotspot?: SanityImageHotspot;
    }>
  | SanityKeyed<Video>
  | SanityKeyed<Embed>
>;

export type Video = {
  _type: "video";
  /**
   * url — `url`
   *
   *
   */
  url?: string;

  /**
   * placeholder — `image`
   *
   *
   */
  placeholder?: {
    _type: "image";
    asset: SanityReference<SanityImageAsset>;
    crop?: SanityImageCrop;
    hotspot?: SanityImageHotspot;
  };

  /**
   * Aspect Ratio — `string`
   *
   * width / height
   */
  aspectRatio?: string;
};

export type Embed = {
  _type: "embed";
  /**
   * url — `url`
   *
   *
   */
  url?: string;
};

export type Contact = {
  _type: "contact";
  /**
   * label — `string`
   *
   *
   */
  label?: string;

  /**
   * value — `string`
   *
   *
   */
  value?: string;

  /**
   * url — `string`
   *
   *
   */
  url?: string;
};

export type TitleText = {
  _type: "titleText";
  /**
   * Title — `string`
   *
   *
   */
  title?: string;

  /**
   * text — `blockContent`
   *
   *
   */
  text?: BlockContent;
};

export type TextUI = {
  _type: "textUI";
  /**
   * title — `string`
   *
   * Module title (displayed only in the admin)
   */
  title?: string;

  /**
   * Text — `blockContent`
   *
   *
   */
  text?: BlockContent;
};

export type ImageUI = {
  _type: "imageUI";
  /**
   * title — `string`
   *
   * Module title (displayed only in the admin)
   */
  title?: string;

  /**
   * Image — `image`
   *
   *
   */
  image?: {
    _type: "image";
    asset: SanityReference<SanityImageAsset>;
    crop?: SanityImageCrop;
    hotspot?: SanityImageHotspot;
  };

  /**
   * width — `number`
   *
   * in percentage (default 100)
   */
  width?: number;

  /**
   * Image caption — `text`
   *
   *
   */
  caption?: string;
};

export type ImagesUI = {
  _type: "imagesUI";
  /**
   * title — `string`
   *
   * Module title (displayed only in the admin)
   */
  title?: string;

  /**
   * images — `array`
   *
   *
   */
  images?: Array<SanityKeyed<ImageUI>>;

  /**
   * width — `number`
   *
   * in percentage
   */
  idth?: number;

  /**
   * Image caption — `text`
   *
   *
   */
  caption?: string;
};

export type TextImageUI = {
  _type: "textImageUI";
  /**
   * title — `string`
   *
   * Module title (displayed only in the admin)
   */
  title?: string;

  /**
   * Text — `blockContent`
   *
   *
   */
  text?: BlockContent;

  /**
   * image — `image`
   *
   *
   */
  image?: {
    _type: "image";
    asset: SanityReference<SanityImageAsset>;
    crop?: SanityImageCrop;
    hotspot?: SanityImageHotspot;
  };
};

export type EmbedUI = {
  _type: "embedUI";
  /**
   * Title — `string`
   *
   * Title
   */
  title?: string;

  /**
   * video — `video`
   *
   *
   */
  video?: Video;

  /**
   * width — `number`
   *
   * in percentage (default 100)
   */
  width?: number;
};